Move snapshot entities into the separate crate snapshot-lib
