+++ catalyst-toolbox/src/vote_check/explorer.rs
@@ -2,8 +2,8 @@ use graphql_client::GraphQLQuery;
 
 #[derive(GraphQLQuery)]
 #[graphql(
-    query_path = "resources/explorer/transaction_by_id.graphql",
-    schema_path = "resources/explorer/schema.graphql",
+    query_path = "../resources/explorer/transaction_by_id.graphql",
+    schema_path = "../resources/explorer/schema.graphql",
     response_derives = "Debug"
 )]
 pub struct TransactionById;
