+++ resources/explorer/transaction_by_id.graphql
