+++ snapshot-lib/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "voting-hir"
+name = "snapshot-lib"
 version = "0.1.0"
 edition = "2021"
 
@@ -13,6 +13,21 @@ jormungandr-lib = { git = "https://github.com/input-output-hk/jormungandr.git", 
 serde = { version = "1", features = ["derive"], optional = true }
 proptest = { git = "https://github.com/input-output-hk/proptest.git", branch = "master", optional = true }
 hex = { version = "0.4", optional = true }
+thiserror = "1.0"
+fraction = "0.10"
+reqwest = { version = "0.11", features = ["blocking", "json"] }
+bech32 = "0.8.1"
+graphql_client = { version = "0.10" }
+chain-crypto = { git = "https://github.com/input-output-hk/chain-libs.git", branch = "master" }
+rust_decimal = "1.16"
+rust_decimal_macros = "1"
+
+[dev-dependencies]
+serde_test = "1"
+test-strategy = "0.2"
+serde_json = "1.0"
+serde_yaml = "0.8.17"
+chain-addr = { git = "https://github.com/input-output-hk/chain-libs.git", branch = "master" }
 
 [features]
 proptest = ["dep:proptest", "hex"]
