+++ snapshot-lib/src/registration.rs
