+++ snapshot-lib/src/voter_hir.rs
@@ -2,9 +2,6 @@
 use ::serde::{Deserialize, Serialize};
 use jormungandr_lib::{crypto::account::Identifier, interfaces::Value};
 
-#[cfg(feature = "proptest")]
-pub mod proptest;
-
 pub type VotingGroup = String;
 
 /// Define High Level Intermediate Representation (HIR) for voting
@@ -49,5 +46,47 @@ mod serde {
         let hex = String::deserialize(deserializer)?;
         Identifier::from_hex(hex.trim_start_matches("0x"))
             .map_err(|e| D::Error::custom(format!("invalid public key: {}", e)))
+    }
+}
+
+#[cfg(any(test, feature = "proptest"))]
+pub mod tests {
+    use super::*;
+    use ::proptest::{prelude::*, strategy::BoxedStrategy};
+    use jormungandr_lib::crypto::account::Identifier;
+    use std::ops::Range;
+
+    impl Arbitrary for VoterHIR {
+        type Parameters = (String, VpRange);
+        type Strategy = BoxedStrategy<Self>;
+        fn arbitrary_with(args: Self::Parameters) -> Self::Strategy {
+            (any::<([u8; 32])>(), args.1 .0)
+                .prop_map(move |(key, voting_power)| VoterHIR {
+                    voting_key: Identifier::from_hex(&hex::encode(key)).unwrap(),
+                    voting_power: voting_power.into(),
+                    voting_group: args.0.clone(),
+                })
+                .boxed()
+        }
+    }
+
+    pub struct VpRange(Range<u64>);
+
+    impl VpRange {
+        pub const fn ada_distribution() -> Self {
+            Self(1..45_000_000_000)
+        }
+    }
+
+    impl Default for VpRange {
+        fn default() -> Self {
+            Self(0..u64::MAX)
+        }
+    }
+
+    impl From<Range<u64>> for VpRange {
+        fn from(range: Range<u64>) -> Self {
+            Self(range)
+        }
     }
 }
